Revlib Implementation (#5)
* Update MotorOutputs

I didn't realize that records are stupid and immutable, so this fixes
that issue

* Add REVLib vendordep

* Add SparkIO

Definitely not done with this.
I'm just pushing the code to have something to look at and ask about
This definitely doesn't feel super clean, but I don't know what to do
about it.

* Update motor outputs

Added supply voltage to the outputs too
I'm starting to get unsure about how I feel about this because it is
starting to feel like a lot, especially if it is for every motor, but
whatever

* Remove SparkIO class

It is just kind of better to separate the two classes, especially for
simulation stuff in the future

* Add spark flex code

Yes, it is exactly the same as the spark max, but it is fine

* Rename spark controller to exploded

Basically, I am taking the internals and making them visible without
requiring a method call first...
I'm exploding the spark...
And because it is a private static class, you can see exactly what it is
exploding

* Add documentation