Add logging to motor component
I don't know what else we are going to be logging, but I left room for
it to be logged.
This is not like the DigitalIO where we are generally only going to
care about the two outputs, here we may have multiple outputs depending
on what the component actually is, so here we are.
A base level output of the motor into a subfolder to leave the rest of
the data free in case of more logging